+++ .config/beemo/prettier.ts
@@ -2,6 +2,8 @@ import { PrettierConfig } from '@beemo/driver-prettier';
 
 const config: PrettierConfig = {
   ignore: ['deno'],
+  useTabs: false,
+  tabWidth: 2,
 };
 
 export default config;