+++ packages/plugin-sub-graph/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### 2.3.0-alpha.1 - 2021-06-28
+
+**Note:** Version bump only for package @giraphql/plugin-sub-graph
+
+
+
+
+
+## 2.3.0-alpha.0 - 2021-06-28
+
+#### 🚀 Updates
+
+- add errors plugin ([88509b4](https://github.com/hayes/giraphql/commit/88509b4))
+
+**Note:** Version bump only for package @giraphql/plugin-sub-graph
+
+
+
+
+
 ### 2.2.2 - 2021-06-11
 
 #### 📦 Dependencies